requirements.txt
Flask==3.1.0
faiss-cpu==1.10.0
sentence-transformers==4.0.2
google-cloud-storage==2.19.0
PyYAML==6.0.2
numpy==1.25.2
scikit-learn==1.6.1
torch==2.6.0
transformers==4.50.3
tqdm==4.67.1